Add zl_picky_t type
As of now, more function pointers could be added from time to time.
It's really annoying that all the time, we have to take care of the type,
which we would like to add and use for functions, also pointers to them.

zl_picky_t type is such a trick for the case, one void pointer,
which is used for taking care of an 'address', and others, pointers to functions,
are packed into with 'union' type. With this mechanism, it can have any types of
function pointers.

That's why it has such name, 'picky'.